address some sonar warnings

+++ src/main/java/org/broadinstitute/consent/http/models/passport/Visa.java
@@ -1,4 +1,4 @@
 package org.broadinstitute.consent.http.models.passport;
 
-public record Visa(String iss, String sub, Integer iat, Integer exp, VisaClaim ga4gh_visa_v1) {
+public record Visa(String iss, String sub, Long iat, Long exp, VisaClaim ga4gh_visa_v1) {
 }
